backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
ifcopenshell==0.8.3.post1
isodate==0.7.2
lark==1.2.2
numpy==2.3.1
packaging==25.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
setuptools==80.9.0
setuptools-scm==8.3.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
